removed nock and unecessary files
