--- helmfile-dependency-check/__test__/test-data/helmfile-lock-stale/index.yaml
@@ -1,77 +1,0 @@
-apiVersion: v1
-entries:
-  datadog:
-    - apiVersion: v1
-      appVersion: "7"
-      created: "2020-11-09T13:34:31.121372087Z"
-      dependencies:
-        - condition: datadog.kubeStateMetricsEnabled
-          name: kube-state-metrics
-          repository: https://charts.helm.sh/stable
-          version: =2.8.11
-      description: Datadog Agent
-      digest: 78f6be3a2aed7a3b2511c6bf99a36d4ac453d507c916746e140ce2cc2e95ab45
-      home: https://www.datadoghq.com
-      icon: https://datadog-live.imgix.net/img/dd_logo_70x75.png
-      keywords:
-        - monitoring
-        - alerting
-        - metric
-      maintainers:
-        - email: support@datadoghq.com
-          name: Datadog
-      name: datadog
-      sources:
-        - https://app.datadoghq.com/account/settings#agent/kubernetes
-        - https://github.com/DataDog/datadog-agent
-      urls:
-        - https://github.com/DataDog/helm-charts/releases/download/datadog-2.5.1/datadog-2.5.1.tgz
-      version: 2.5.1
-    - apiVersion: v1
-      appVersion: "7"
-      created: "2020-11-06T16:59:54.198467615Z"
-      dependencies:
-        - condition: datadog.kubeStateMetricsEnabled
-          name: kube-state-metrics
-          repository: https://charts.helm.sh/stable
-          version: =2.8.11
-      description: Datadog Agent
-      digest: a9f33a1f5404788904ad12e4869d0115071b9f0e8fe56c25ab7764a8fcd679f3
-      home: https://www.datadoghq.com
-      icon: https://datadog-live.imgix.net/img/dd_logo_70x75.png
-      keywords:
-        - monitoring
-        - alerting
-        - metric
-      maintainers:
-        - email: support@datadoghq.com
-          name: Datadog
-      name: datadog
-      sources:
-        - https://app.datadoghq.com/account/settings#agent/kubernetes
-        - https://github.com/DataDog/datadog-agent
-      urls:
-        - https://github.com/DataDog/helm-charts/releases/download/datadog-2.5.0/datadog-2.5.0.tgz
-      version: 2.5.0
-  synthetics-private-location:
-    - apiVersion: v2
-      appVersion: 1.4.0
-      created: "2020-10-27T13:52:42.263511255Z"
-      description: Datadog Synthetics Private Location
-      digest: ae8714fd2e404f25456b21eceb5d13b229bd4b9e154a97ceed69a7c174e16d43
-      home: https://www.datadoghq.com
-      icon: https://datadog-live.imgix.net/img/dd_logo_70x75.png
-      keywords:
-        - monitoring
-        - synthetics
-      maintainers:
-        - email: support@datadoghq.com
-          name: Datadog
-      name: synthetics-private-location
-      sources:
-        - https://docs.datadoghq.com/synthetics/private_locations
-        - https://app.datadoghq.com/synthetics/settings/private-locations
-      urls:
-        - https://github.com/DataDog/helm-charts/releases/download/synthetics-private-location-0.1.0/synthetics-private-location-0.1.0.tgz
-      version: 0.1.0
-generated: "2020-07-29T11:03:41.153796856Z"
