+++ package.json
@@ -18,10 +18,8 @@
     "@types/jest": "^26.0.15",
     "@types/js-yaml": "^3.12.5",
     "@types/minimatch": "^3.0.3",
-    "@types/nock": "^11.1.0",
     "@vercel/ncc": "^0.25.1",
     "jest": "^26.6.3",
-    "nock": "^13.0.5",
     "npm-run-all": "^4.1.5",
     "ts-jest": "^26.4.4",
     "typescript": "^4.0.5"
